[+:] [Address iterator] Added GetNextAddr() shortcut..
"return GetAddr(Address) && AdvanceAddr();"

So now it's not purely abstract class. But in most cases your iteration
is "while (Rator.GetNextAddr(&Addr))". Only in rare cases you want
to get address without advancing it.

So without this shortcut code, interface is not practical. Users will
use implementations with GetNextAddr() and pass them as parameters.
That's completely defeats idea of interface: pass only core info.

This change increases compiled code size. We're not happy but accept it.